move ontologies data collation from explorer component to a service
manage-explorer.js : move functions :

to utils/data/block-values.js
  blockValues
  blockValuesBlocks
  blockValuesViewed
  storeBlockAttributes
  checkPositions
  blockValuesHistory
  addField
  blocksFilterCurrentlyViewed
  blockValuesIdText
  blockValuesNameFiltered
  blockValuesTree
  ids2Blocks
  blockFeatureOntologiesTreeEmbeddedFn
  idParentNodeMap

to utils/value-tree.js :
  collateOntologyId2Node
  walkTree
  mapTree0
  mapTree
  treeFor
  copyNode
  addNode
  treesChildrenCount

to services/data/block-values.js :
  blockFeatureOntologies
  blockFeatureOntologiesBlocks
  ontologyId2Node
  ontologyId2NodeFor
  levelMeta

add services/data/block-values.js, with these CPs copied from manage-explorer.js :

blockFeatureOntologiesViewed	(added, based on blockFeatureOntologiesBlocks)
blockFeatureOntologiesTree
blockFeatureOntologiesViewedEmbedded (partly copied from blockFeatureOntologiesTreeEmbedded)
blockFeatureOntologiesTreeEmbedded
ontologyId2DatasetNodes

utils/data/grouping.js : factor out blocksValuesUnwindAndGroup() from blocksParentAndScope().

in manage-explorer.js, these moved CPs are replaced by aliases to blockValues. :
  blockFeatureOntologies
  blockFeatureOntologiesBlocks
  ontologyId2Node
  ontologyId2NodeFor
  levelMeta

utils/value-tree.js : add blockValuesViewed()

blockValuesHistory() : add param historyView
blockFeatureOntologiesTree() : add param valueName
mapTree() : check if tree.id is defined.
