+++ frontend/app/components/panel/manage-explorer.js
@@ -4,7 +4,6 @@ import { resolve, all } from 'rsvp';
 import { A } from '@ember/array';
 import { inject as service } from '@ember/service';
 import { isArray } from '@ember/array';
-import { singularize, pluralize } from 'ember-inflector';
 
 
 import DS from 'ember-data';
@@ -35,10 +34,23 @@ import { tab_explorer_prefix, text2EltId } from '../../utils/explorer-tabId';
 import { parseOptions } from '../../utils/common/strings';
 import { thenOrNow } from '../../utils/common/promises';
 
-import { valueGetType, mapHash, reduceHash, reduceIdChildrenTree, justUnmatched,
-         logV, ontologyIdFromIdText } from '../../utils/value-tree';
+import {
+  valueGetType, mapHash, reduceHash, reduceIdChildrenTree, justUnmatched,
+  logV,
+  ontologyIdFromIdText,
+  treeFor, treesChildrenCount,
+ } from '../../utils/value-tree';
 import { blocksParentAndScope } from '../../utils/data/grouping';
 
+import {
+  blockValues,
+  blockValuesBlocks,
+  blockValuesHistory,
+  blockValuesNameFiltered,
+  blockValuesTree,
+  blockFeatureOntologiesTreeEmbeddedFn,
+  idParentNodeMap,
+} from '../../utils/data/block-values';
 
 import ManageBase from './manage-base'
 
@@ -68,369 +80,7 @@ const dLog = console.debug;
 const selectorExplorer = 'div#left-panel-explorer';
 
 
-/*----------------------------------------------------------------------------*/
 
-/**
- * CP : blockFeatureTraits
- * Used in blockFeature{Traits,Ontologies}, which are used only for checking
- * .length in panel/{manage-explorer,ontologies}.hbs
- * @param fieldName 'Traits' or 'Ontologies'
- */
-function blockValues(fieldName) {
-  let ObjectPromiseProxy = ObjectProxy.extend(PromiseProxyMixin);
-
-  let count = this.get('blocksService.featureUpdateCount');
-  dLog('blockValues', fieldName, count, this);
-  let valueP = this.get('apiServerSelectedOrPrimary.blockFeature' + fieldName);
-  let proxy = ObjectPromiseProxy.create({ promise: resolve(valueP) });
-  return proxy;
-}
-
-/** map ._id to .block
- * CP : blockFeatureTraitsBlocks
- * Used by : blockValuesHistory()
- * @param fieldName 'Traits' or 'Ontologies'
- */
-function blockValuesBlocks(fieldName) {
-  // 'blockFeatureTraits'
-  let apiServer = this.get('apiServerSelectedOrPrimary');
-  let blocksTraitsP = apiServer.get('blockFeature' + fieldName);
-  let store = apiServer.get('store');
-  /** ids2Blocks() depends on this result. */
-  if (! apiServer.get('datasetsBlocks')) {
-    blocksTraitsP = Promise.resolve([]);
-  } else {
-    blocksTraitsP = blocksTraitsP
-      .then((blocksTraits) => {
-        blocksTraits = ids2Blocks(store, blocksTraits);
-        blocksTraits = checkPositions(blocksTraits, fieldName);
-        storeBlockAttributes(blocksTraits, fieldName);
-        return blocksTraits;
-      });
-  }
-  return blocksTraitsP;
-}
-
-/** Store the Traits / Ontologies of each block, to enable loadBlock() to make
- * them visible.
- * @param blocksValues is the result of checkPositions(), i.e. QTLs which will
- * not be displayed because of lack of position / .values.<fieldName> are filtered out.
- */
-function storeBlockAttributes(blocksValues, fieldName) {
-  blocksValues.forEach((bt) => {
-    let
-    block = bt.block,
-    values = bt[fieldName],
-    attr = block.get('attributes') || block.set('attributes', {});
-    attr[fieldName] = values;
-  });
-}
-
-/** map blocksTraits : if .block has .positioned.<fieldName>
- * use that in place of .<fieldName>
- * @param fieldName is plural; the singular form is used for .positioned.<fieldName>
- */
-function checkPositions(blocksTraits, fieldName) {
-  let fieldNameSingular = singularize(fieldName);
-  let bts = blocksTraits.map((bt) => {
-    let names = bt.block.get('positioned.' + fieldNameSingular);
-    if (names && (names.length !== bt[fieldName].length)) {
-      dLog('checkPositions', bt[fieldName], names);
-      bt = Object.assign({}, bt);
-      bt[fieldName] = names;
-    }
-    return bt;
-  });
-  return bts;
-}
-
-/**
- * CP : blockFeatureTraitsHistory
- */
-function blockValuesHistory (fieldName) {
-  let blocksTraitsP = this.get('blockFeature' + fieldName + 'Blocks');
-  if (this.historyView !== 'Normal') {
-    blocksTraitsP = blocksTraitsP
-      .then((blocksTraits) => {
-        const
-        recent = this.historyView === 'Recent',
-        /** map blocks -> Traits, so that the sorted blocks can be mapped -> blocksTraits  */
-        blocksTraitsMap = blocksTraits.reduce((btm, bt) => btm.set(bt.block, bt[fieldName]), new Map()),
-        blocks = blocksTraits.map((bt) => bt.block),
-        /** sorted blocks */
-        blocksS = (this.historyView === 'Viewed') ?
-          blocksFilterCurrentlyViewed(blocks) :
-          this.get('viewHistory').blocksFilterSortViewed(blocks, recent),
-        blocksTraitsS = blocksS.map((b) => addField({block : b}, fieldName, blocksTraitsMap.get(b)));
-        return blocksTraitsS;
-      });
-  }
-  return blocksTraitsP;
-}
-
-
-function addField(object, fieldName, value) {
-  object[fieldName] = value;
-  return object;
-}
-
-/** Filter the given array of blocks to just those which are currently viewed.
-*/
-function blocksFilterCurrentlyViewed(blocks) {
-  let
-  blocksSorted = blocks
-    .filter((b) => b.isViewed);
-  return blocksSorted;
-};
-
-
-/**
- * Used as a pre-process for (fieldName === 'Ontologies')
- * in blockValuesNameFiltered (CP : blockFeatureOntologiesName)
- * @param me  manage-explorer
- */
-function blockValuesIdText(me, blocksTraits) {
-  blocksTraits.forEach((bt) => {
-    bt.Ontologies = bt.Ontologies
-      .filter((oid) => oid !== '')
-      .map((oid)=> {
-        let result = oid;
-        if (! oid.startsWith('[')) {
-          let name = me.get('ontology').getNameViaPretzelServer(oid);
-          if (typeof name === 'string') {
-            result = '[' + oid + '] ' + name;
-          }
-        }
-        return result;
-      });
-  });
-  return blocksTraits;
-}
-
-
-/**
- * CP : blockFeatureTraitsName
- */
-function blockValuesNameFiltered (fieldName) {
-  let
-  nameFilters = this.get('nameFilterArray'),
-  blocksTraitsP = this.get('blockFeature' + fieldName + 'History');
-
-  if (fieldName === 'Ontologies') {
-    blocksTraitsP = blocksTraitsP
-      .then((bts) => blockValuesIdText(this, bts) );
-  }
-
-  if (nameFilters.length) {
-    blocksTraitsP = blocksTraitsP
-      .then((blocksTraits) => {
-        blocksTraits = blocksTraits
-          .map((blockTraits) => this.blockTraitsFilter(fieldName, blockTraits, nameFilters))
-          .filter((blockTraits) => blockTraits[fieldName].length);
-        return blocksTraits;
-      });
-  }
-  return blocksTraitsP;
-}
-
-/*
- * CP : blockFeatureTraitsTree
- */
-function blockValuesTree (fieldName) {
-  let ObjectPromiseProxy = ObjectProxy.extend(PromiseProxyMixin);
-
-  let
-  valueP = this.get('blockFeature' + fieldName + 'Name')
-    .then((blocksTraits) => {
-      let
-      blocksTraitsTree = blocksParentAndScope(this.get('levelMeta'), fieldName, blocksTraits);
-      this.set('blockFeature' + fieldName + 'TreeKeyLength', Object.keys(blocksTraitsTree).length);
-      return blocksTraitsTree;
-    });
-  let proxy = ObjectPromiseProxy.create({ promise: resolve(valueP) });
-
-  return proxy;
-}
-
-/** map blockIdsTraits[]
- * from {_id, Traits} to {block, Traits}
- * or from {_id, Ontologies} to {block, Ontologies}
- */
-function ids2Blocks(store, blockIdsTraits) {
-  let
-  blocksTraits = store && blockIdsTraits
-    .map(({_id, ...rest}) => (rest.block = store.peekRecord('block', _id), rest))
-    .filter((bt) => bt.block);
-  return blocksTraits;
-}
-
-
-/** reduce a tree to a result.
- * Used for Ontology tree, result of ontology service getTree();
- * each tree node has .id and .children.
- * Added : incorporate multiple roots into a single tree, so the top level in
- * that case will be an object { <rootId> : tree, ... }.
- * Similar signature to Array.reduce().
- * @param result
- * @param reduceNode Add a node to a result function reduceNode(result, node) -> result;
- * @param tree
- */
-function walkTree(result, reduceNode, tree) {
-  const fnName = 'walkTree';
-  result = reduceNode(result, tree);
-  let children = tree.children;
-  if (isArray(children)) {
-    result = children.reduce((result, node) => {
-      result = walkTree(result, reduceNode, node);
-      return result;
-    }, result);
-  }
-  // dLog(fnName, result, tree);
-  return result;
-};
-
-/** Similar to mapTree(), this produces a tree of nodes which are each just {id : value},
- * whereas mapTree() allows {id, type, node} to be included in the node.
- */
-function mapTree0(levelMeta, result, tree) {
-  let value = result[tree.id] = {};
-  levelMeta.set(value, tree.type);
-  tree.children.forEach((c) => mapTree0(levelMeta, value, c));
-  return result;
-};
-/** Map the Ontology tree to a value-tree, with nodes containing :
- *   {id, type, children, node : dataset_ontology}
- * where dataset_ontology links to the corresponding node in the tree of
- * datasets by Ontology.
- * @param id2Node map OntologyId to {<parent> : {<scope> : [block, ...], ...}, ...}
- * @param tree  Ontology API result
- */
-function mapTree(levelMeta, id2Node, tree) {
-  let
-  /** used by copyNode() to lookup a value for .parent */
-  id2nc = {},
-  value = copyNode(levelMeta, id2nc, tree),
-
-  node = id2Node[tree.id];
-  if (node) {
-    value.node = node;
-    dLog('mapTree', value, node);
-  }
-  if (value.text || value.id) {
-    /** same format as ontologyNameId(), rootOntologyNameId()     */
-    value.name = '[' + value.id + ']  ' + value.text;
-  }
-
-  if (tree.children === undefined) {
-    /* copyNode copies attributes for node with .id & .children, but not for plain object. */
-    /* optional : id2nc[value.name] = value;
-     * copyNode() does id2nc[n.id] = c, i.e. id2nc[value.id] = value, but value.id is undefined
-     */
-    Object.entries(tree)
-      .reduce((result, e) => {
-        result[e[0]] = mapTree(levelMeta, id2Node, e[1]);
-        return result;
-      }, value);
-  } else {
-    let children = tree.children;
-    if (isArray(children)) {
-      value.children = children.map((c) => {
-        /** copy of c */
-        let cc = mapTree(levelMeta, id2Node, c);
-        cc.parent = value;
-        return cc; });
-    } else if (children.children && children.id && children.text && children.type)
-    {
-      dLog('mapTree', 'value', value, 'tree', tree);      
-      tree.children = [tree.children];
-    }
-  }
-
-  return value;
-};
-
-/*----------------------------------------------------------------------------*/
-
-/** Make a copy of tree, which is addressed by id2n, with only the branches required
- * to support id2Pn.
- * @param levelMeta to record the node types of the output tree
- * @param tree  ontologyTree
- * @param id2n  ontologyId2Node : references from OntologyId into the corresponding nodes in ontologyTree
- * @param id2Pn ontologyId2DatasetNodes : references from OntologyId into the parent nodes of blockFeatureOntologiesTree
- */
-function treeFor(levelMeta, tree, id2n, id2Pn) {
-  let
-  id2nc = {},
-  treeCopy = copyNode(levelMeta, id2nc, tree);
-  /* could use forEachHash(), or pass {id2n, id2nc} as result in & out;
-   * or filterHash() | mapHash() (will that add the root ok ?).
-   */
-  /*treeCopy =*/ reduceHash(id2Pn, (t, oid, p) => {
-    oid = ontologyIdFromIdText(oid);
-    let on = id2n[oid];
-    if (! on) {
-      dLog('treeCopy', oid, 'not present in', id2n);
-    } else {
-      addNode(levelMeta, id2nc, on);
-    }
-  }, treeCopy);
-  return treeCopy;
-};
-/** Copy just 1 node. */
-function copyNode(levelMeta, id2nc, n) {
-  /** copy of node n */
-  let c;
-  if (n.children === undefined) {
-    dLog('copyNode', 'roots', n);
-    c = {};
-  } else {
-    c = Object.assign({}, n);
-    if (typeof n.children !== 'boolean') {
-      c.children = [];
-    }
-  }
-  levelMeta.set(c, n.type);
-  id2nc[n.id] = c;
-  if (n.parent) {
-    c.parent = id2nc[n.parent.id];
-  }
-  return c;
-}
-/** Add a copy of on to t, and supporting branch.
- * @param id2nc index into treeCopy.
- * @param on  node from ontologyTree
- */
-function addNode(levelMeta, id2nc, on) {
-  let onc = id2nc[on.id];
-  if (! onc) {
-    let parent = on.parent;
-    if (parent) {
-      // change parent to refer to the copy of parent.
-      parent = addNode(levelMeta, id2nc, parent);
-    }
-    onc = copyNode(levelMeta, id2nc, on);
-    if (parent) {
-      if (parent.children) {
-        parent.children.push(onc);
-      } else {
-        parent[onc.id] = onc;
-      }
-    }
-  }
-  return onc;
-}
-
-/*----------------------------------------------------------------------------*/
-
-/** @return a count of the children at the top level, or 2nd level if valueTree is multiple roots.
- * @desc possibly displaying the leaf count would be more useful.
- */
-function treesChildrenCount(valueTree) {
-  let count = valueTree.children ?
-      valueTree.children.length :
-      Object.values(valueTree).reduce((result, vt) => result += vt.children.length, 0);
-  return count;
-}
 /*----------------------------------------------------------------------------*/
 
 
@@ -441,9 +91,11 @@ export default ManageBase.extend({
   blocksService : service('data/block'),
   ontology : service('data/ontology'),
   trait : service('data/trait'),
+  blockValues : service('data/block-values'),
 
   init() {
-    this._super();
+    this._super(...arguments);
+
     if (initRecursionCount++ > 5) {
       debugger;
       window.alert('initRecursionCount : ' + initRecursionCount);
@@ -454,7 +106,6 @@ export default ManageBase.extend({
     /** Initialise this.blocksService so that dependency blocksService.featureUpdateCount works. */
     let blocksService = this.get('blocksService');
 
-    this.treeFor = treeFor; // will move to utils/data/grouping.js
     this.get('ontology').set('ontologyCollation', this);
   },
 
@@ -556,7 +207,7 @@ export default ManageBase.extend({
 
   blockFeatureTraitsHistory : computed(
     'blockFeatureTraitsBlocks', 'historyView',
-    function () { return blockValuesHistory.apply(this, ['Traits']); }),
+    function () { return blockValuesHistory.apply(this, ['Traits', this.historyView]); }),
 
   blockFeatureTraitsName : computed(
     'blockFeatureTraitsHistory.[]',
@@ -565,29 +216,19 @@ export default ManageBase.extend({
 
   blockFeatureTraitsTree : computed(
     'blockFeatureTraitsName',
-    function () { return blockValuesTree.apply(this, ['Traits']); }),
+    function () { return blockValuesTree.apply(this, ['Traits', 'blockFeatureTraitsName']); }),
 
 
   /*--------------------------------------------------------------------------*/
   /** Implement Ontology tab, as for Trait tab */
 
-  blockFeatureOntologies : computed(
-    'apiServerSelectedOrPrimary.blockFeatureOntologies',
-    'apiServerSelectedOrPrimary.datasetsBlocks.[]',
-    /** comment in feature-edit.js : saveFeature() */
-    'blocksService.featureUpdateCount',
-    function () { return blockValues.apply(this, ['Ontologies']); }),
+  blockFeatureOntologies : alias('blockValues.blockFeatureOntologies'),
+  blockFeatureOntologiesBlocks : alias('blockValues.blockFeatureOntologiesBlocks'),
 
-  /** map ._id to .block */
-  blockFeatureOntologiesBlocks : computed(
-    'apiServerSelectedOrPrimary.blockFeatureOntologies',
-    'apiServerSelectedOrPrimary.datasetsBlocks.[]',
-    'blocksService.featureUpdateCount',
-    function () { return blockValuesBlocks.apply(this, ['Ontologies']); }),
 
   blockFeatureOntologiesHistory : computed(
     'blockFeatureOntologiesBlocks', 'historyView',
-    function () { return blockValuesHistory.apply(this, ['Ontologies']); }),
+    function () { return blockValuesHistory.apply(this, ['Ontologies', this.historyView]); }),
 
   blockFeatureOntologiesName : computed(
     'blockFeatureOntologiesHistory.[]',
@@ -598,7 +239,7 @@ export default ManageBase.extend({
 
   blockFeatureOntologiesTree : computed(
     'blockFeatureOntologiesName',
-    function () { return blockValuesTree.apply(this, ['Ontologies']); }),
+    function () { return blockValuesTree.apply(this, ['Ontologies', 'blockFeatureOntologiesName']); }),
 
   /** used when .showHierarchy === true
    */
@@ -642,22 +283,10 @@ export default ManageBase.extend({
     treeP = this.get('treesForData'), // single-root : ontologyTree
     id2PnP = this.get('ontologyId2DatasetNodes'),
     promise = Promise.all([treeP, id2PnP]).then(([tree, id2Pn]) => {
-      dLog(fnName, 'id2Pn', id2Pn);
-      let
-      valueTree = mapTree(this.get('levelMeta'), id2Pn, tree);
-      this.levelMeta.set(valueTree, 'term');
-
+      let valueTree = blockFeatureOntologiesTreeEmbeddedFn(this.levelMeta, tree, id2Pn);
       /** valueTree is the root, e.g.  "CO_321:ROOT" : "Wheat traits", so count the children. */
       let keyLength = treesChildrenCount(valueTree);
       this.set('blockFeatureOntologiesTreeEmbeddedKeyLength', keyLength);  // perhaps rename both to keysLength.
-
-      Object.values(valueTree).forEach((t) => t.parent = valueTree);
-      /*
-      let childNames = Object.keys(valueTree);
-      valueTree.name = childNames.length ? childNames[0].slice(0,2) : 'CropOntology';
-      */
-
-      dLog('blockFeatureOntologiesTreeEmbedded', valueTree);
       return valueTree;
     });
     return promise;
@@ -724,19 +353,6 @@ export default ManageBase.extend({
     promise = blocksOntologiesTree
       .then((bot) => idParentNodeMap(bot));
 
-    function idParentNodeMap(bot) {
-      /** traverse the parent level, add the Ontology ID Node {id, blocks} */
-      let id2n = reduceHash(
-        bot,
-        (result, key, value) => {
-          key = ontologyIdFromIdText(key);
-          (result[key] ||= []).push(value);
-          return result;
-        },
-        {});
-      return id2n;
-    };
-
     return promise;
   }),
 
@@ -746,33 +362,9 @@ export default ManageBase.extend({
     let treeP = this.get('ontology').getTree();
     return treeP;
   }),
-  /** collate a mapping [OntologyId] -> tree node */
-  ontologyId2Node : computed('blockFeatureOntologiesTreeEmbedded', function () {
-    /** use blockFeatureOntologiesTreeEmbedded in place of ontologyTree, to have .parent. */
-    let treeP = this.get('blockFeatureOntologiesTreeEmbedded');
-    let id2node = treeP.then((tree) => this.collateOntologyId2Node(tree));
-    return id2node;
-  }),
-  /** similar to ontologyId2Node, based on blockFeatureOntologiesTreeGrouped
-   * which is filtered according to explorer filter and Viewed.  */
-  ontologyId2NodeFor : computed('blockFeatureOntologiesTreeGrouped', function () {
-    let treeP = this.get('blockFeatureOntologiesTreeGrouped');
-    let id2node = treeP.then((tree) => this.collateOntologyId2Node(tree));
-    return id2node;
-  }),
-  collateOntologyId2Node(tree) {
-    /** Add a node to a result. */
-    function reduceNode(result, parentKey, index, node) {
-      let id = node.id;
-      if (id) {
-        result[id] = node;
-      }
-      return result;
-    }
-    let
-    id2node = reduceIdChildrenTree(tree, reduceNode, {});
-    return id2node;
-  },
+
+  ontologyId2Node : alias('blockValues.ontologyId2Node'),
+  ontologyId2NodeFor : alias('blockValues.ontologyId2NodeFor'),
 
   /*--------------------------------------------------------------------------*/
 
@@ -1213,7 +805,7 @@ export default ManageBase.extend({
 
   /* ------------------------------------------------------------------------ */
 
-  levelMeta : new WeakMap(),
+  levelMeta : alias('blockValues.levelMeta'),
 
   /** group all datasets by parent type
    * add tabs for those, FG can apply, so:
